Add license, haxelib.json, and basic class layout.
An `Object` is the base framework element. It handles the flags that make
it capable of partaking in the framework game life cycle.

`Entity` is an extension of `Object` that provides position and size
properties, as well as a hitbox and collision flag.

`Sprite` is an extension of `Entity` that currently differs only in
appearance since it draws an image to the framebuffer.

Collision is handled through shapes. The base `Shape` class only provides
a position because that is a requirement for any shape. `Rectangle` thus
extends shape with a width and height property as well as collision logic,
which will differ from shape to shape.