Add `Game` and `State` types, globals (`Khake`).
A `State` is a `Group`. Currently it is scant, defining only overrideable `onCreate()` and `onDestroy()` methods (the latter isn't even used yet).

`Game` prepares a backbuffer in accordance with provided width and
height. It handles states and state changes, and makes sure the currently
active state is updated and drawn.

`Khake` is intended to contain globals. Currently this is only a reference
to the `Game` object.