Add `Engine` class, modify `Game`, update comments
The engine serves as a way to reduce Kha's admittedly minimal boilerplate.
As arguments it takes a `khake.Game` class (which it later instantiates),
the window size, and the desired update frame rate. It also takes care of
tracking the elapsed time and delta time (time that passes between
frames). Relevant information is stored in `Khake` globals.

The game was modified accordingly. Some of the `khake.Game` class
properties were migrated to `Engine` (such as `frameRate`), and some
properties were exposed to the end user via `Khake` globals.

The rest of the codebase received updates to the comment style.