Give more freedom to Engine, Game classes
The engine now mostly cares only about updating and rendering the game.

It's up to the game to initialise and update input managers, handle
graphical fidelity, etc. This is possible through the addition of pre- and
post- variants of draw and update functions, as well as some extra
properties regarding the clearing and colouring of back and frame buffers.