Begin restructuring how hitboxes are updated
Instead of relying on update calls, the hitbox should only update when it
is requested by another part of the framework or game logic. Ideally, the
collider should have a reference to at least the position of the entity
the collider's hitbox is attached to. A getter function should then update
the hitbox position according to the reference position before returning
the hitbox to the requester.