Define engine configuration structure
The engine is now initialised and configured through a special
configuration structure. Configuration options are defined in this
structure, which is then passed to the engine's `init()` function.

The changes are reflected in the `Mokha` API, where all global variables
affected by the engine are now accessible for modification only by the
`Engine` class.